feat: move constants for win and unix
