+++ package.json
@@ -20,7 +20,8 @@
     "pkg:win": "yarn prod && pkg -t node14-win dist/verto.js",
     "pkg:arm": "echo not implemented",
     "pkg:all": "yarn prod && pkg dist/verto.js",
-    "test": "rollup -c rollup.test.js && mocha 'build/verto.test.js'"
+    "test": "rollup -c rollup.test.js && mocha 'build/verto.test.js'",
+    "test:integration": "VERTO_URI=http://localhost:3000/verto.zip go run ./test/integration/"
   },
   "dependencies": {
     "@koa/cors": "^3.1.0",
