+++ test/integration/installer.go
@@ -3,10 +3,10 @@ package main
 import "log"
 
 func TestInstaller() {
-	err, stderr, stdout := Shellout("curl -fsSL http://localhost:3000/install/linux.sh | sh")
+	err, stderr, _ := Shellout(InstallCmd)
 	if err != nil {
 		log.Printf("error: %v\n", stderr)
 		panic(err)
 	}
-	log.Println(stdout)
+	Logger.Success(InstallCmd)
 }
