--- test/integration/shell_unix.go
@@ -1,5 +1,0 @@
-// +build !windows
-
-package main
-
-const ShellToUse = "bash"
