--- test/integration/shell_win.go
@@ -1,6 +1,0 @@
-// +build windows
-
-package main
-
-// TODO(littledivy): Use powershell?
-const ShellToUse = "cmd"
